Move all messages and models to the common project
