+++ build.gradle
@@ -50,7 +50,8 @@ subprojects {
     dependencies {
         implementation('org.jetbrains.kotlin:kotlin-reflect')
         implementation('org.jetbrains.kotlin:kotlin-stdlib-jdk8')
-
+        implementation('org.springframework.boot:spring-boot-starter-data-mongodb')
+        
         testImplementation('org.jetbrains.kotlin:kotlin-test-common')
         testImplementation('org.junit.jupiter:junit-jupiter-api:5.2.0')
         testImplementation ('org.spekframework.spek2:spek-dsl-jvm:2.0.0-rc.1')  {
