+++ mb-chatbot/build.gradle
@@ -3,7 +3,6 @@ repositories {
 }
 
 dependencies {
-    implementation(project(':mb-blog'))
     implementation(project(':mb-common'))
 
     implementation('com.linecorp.bot:line-bot-spring-boot:2.4.0')
