+++ mb-chatbot/src/main/kotlin/com/nomkhonwaan/mb/chatbot/line/LINEMessageHandler.kt
@@ -8,8 +8,8 @@ import com.linecorp.bot.model.event.message.TextMessageContent
 import com.linecorp.bot.model.message.Message
 import com.linecorp.bot.spring.boot.annotation.EventMapping
 import com.linecorp.bot.spring.boot.annotation.LineMessageHandler
-import com.nomkhonwaan.mb.blog.post.FindAllPublishedPostsQuery
-import com.nomkhonwaan.mb.blog.post.Post
+import com.nomkhonwaan.mb.common.messaging.post.FindAllPublishedPostsQuery
+import com.nomkhonwaan.mb.common.messaging.post.Post
 import org.axonframework.messaging.responsetypes.ResponseTypes
 import org.axonframework.queryhandling.QueryGateway
