+++ mb-common/build.gradle
@@ -4,4 +4,5 @@ repositories {
 
 dependencies {
     implementation('org.axonframework:axon-spring:4.0.3')
+    implementation('org.springframework.boot:spring-boot-starter-data-mongodb')
 }
