+++ mb-common/src/main/kotlin/com/nomkhonwaan/mb/common/messaging/category/CategoryQueries.kt
@@ -1,4 +1,4 @@
-package com.nomkhonwaan.mb.blog.category
+package com.nomkhonwaan.mb.common.messaging.category
 
 /**
  * Finds all categories Query.
