+++ mb-common/src/main/kotlin/com/nomkhonwaan/mb/common/messaging/post/PostEvents.kt
@@ -1,8 +1,9 @@
-package com.nomkhonwaan.mb.blog.post
+package com.nomkhonwaan.mb.common.messaging.post
 
-import com.nomkhonwaan.mb.blog.category.Category
 import com.nomkhonwaan.mb.common.cqrs.Event
+import com.nomkhonwaan.mb.common.messaging.category.Category
 import java.time.ZonedDateTime
+
 
 /**
  * A Post created Event.
