+++ mb-common/src/main/kotlin/com/nomkhonwaan/mb/common/messaging/user/UserQueries.kt
@@ -1,0 +1,8 @@
+package com.nomkhonwaan.mb.common.messaging.user
+
+/**
+ * Finds a single User by their ID Query.
+ *
+ * @param id An identifier of the User
+ */
+data class FindUserByIDQuery(val id: String)
