+++ mb-graphql/src/main/kotlin/com/nomkhonwaan/mb/graphql/QueryResolver.kt
@@ -1,11 +1,11 @@
 package com.nomkhonwaan.mb.graphql
 
 import com.coxautodev.graphql.tools.GraphQLQueryResolver
-import com.nomkhonwaan.mb.auth.FindUserByIDQuery
-import com.nomkhonwaan.mb.auth.User
-import com.nomkhonwaan.mb.blog.category.Category
-import com.nomkhonwaan.mb.blog.category.FindAllCategoriesQuery
-import com.nomkhonwaan.mb.blog.post.*
+import com.nomkhonwaan.mb.common.messaging.category.Category
+import com.nomkhonwaan.mb.common.messaging.category.FindAllCategoriesQuery
+import com.nomkhonwaan.mb.common.messaging.post.*
+import com.nomkhonwaan.mb.common.messaging.user.FindUserByIDQuery
+import com.nomkhonwaan.mb.common.messaging.user.User
 import org.axonframework.messaging.responsetypes.ResponseTypes
 import org.axonframework.queryhandling.QueryGateway
 import org.springframework.security.access.prepost.PreAuthorize
